# api_gateway/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API task manager docs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "82d74137970376a374510566964c9813",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "email",
                                                "example": "user@gmail.com"
                                            },
                                            "password": {
                                                "type": "password",
                                                "example": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response_code": {
                                            "type": "string",
                                            "example": "200"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success Login"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIxIiwianRpIjoiNGJlMTkxNzBjOWI4ZmQyZGQ5YTZiNjhiOWY0MmMzZjhiZTk1NjQ0ZThiZDEzYTQ0YjU4MjhiZWJhNGQwZDAyN2UzZDI2OWJlMmVhM2IyZmEiLCJpYXQiOjE3Mjg0MDIxNTEuMzEyMjY1LCJuYmYiOjE3Mjg0MDIxNTEuMzEyMjczLCJleHAiOjE3NTk5MzgxNTAuOTMzMzg1LCJzdWIiOiI5Iiwic2NvcGVzIjpbXX0.KpPcP51-kuaermf18uvzNXgohD_Xyo1Q24c421mBHo0HwN3qqDTf6CBB86MKLOKsly2OA3MJCMAOKVBXUoyiIMlcQOHBbneDUe1VRrYzhn6rvZLmxf5cGQYKMfgQTai1bQAAHK1-UNxWlY5AP05pmr00kWC5W8uyUdAQYQA0ZS7DftS_cNOuuAIdonsiEe5EihqA3Zo3KtVqd4PDU1S7ioIFJoABBKU62Z5L239qI5McIFkCoSRzuOzvLCwehE3ZieKh-juLGvlXMemlOzig1dZ1Z8h-pqlBP6dEqRvj8w7t3tkUjCMgLIuazoGdS3-FdfmAFLgJwnBsOkUsxaPHASOhgEScBiCRn51-n2PCtEKgDF_jTOJ_WKsXeXIUWgSbM8KC5DLdzT4htkRGbqXH85LX6WBc381_ovDqyI4Os84Sbwhu7CfA_34BL4XY_sgf1rgkFSaaM-BjwUxy6yjYLIrzKf8dPsCDIELAC75KbvALbP44SoU9zCNYA8kY5U-iivMz3o04ASb6AvddS01eu5-NUnRtXoSZDdCIWX6VjQ7FdTxQ92F0DxUND9CZoQIWGwkXV1HfzbjOlOF9T7_Wul4BFz3hWiDlFddBHcJYfxEDCH9i2oQkoaVQMctN2Y67CeFLcAat6KdFKVkonaUOiGtpfsUCpmOeDIRM-rerVfA"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response_code": {
                                            "type": "string",
                                            "example": "401"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorised"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registration",
                "operationId": "c47411f6de96b5b30018c647e449e470",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "test",
                                                "example": "John"
                                            },
                                            "email": {
                                                "type": "email",
                                                "example": "user@gmail.com"
                                            },
                                            "password": {
                                                "type": "password",
                                                "example": "12345678"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response_code": {
                                            "type": "string",
                                            "example": "200"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success Register"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}